Feat: Refactor Intro Component to be Reusable
Instead of being specific to intro (name, email,
phone number), now the component takes in data
as props to create whatever you pass in.
The data includes text, className, labeltext,
input type (optional).

Used an array of objects to hold data for this
since an array is easily iterable with array methods.