Feat: Allow adding additional education/experience
Did this by holding state of array of ids.
Ids are generated with randomUUID.
Ids are passed as keys into components.

The array is iterated through and render
one Section component per id (with key as that id)

This is ideal because it does not duplicate
data within state (since we are storing ids, not
the data itself again).