Fix: Allow editing of responsibility form values
Previously, the id of the input elements were not
set so it was not possible for the onChange func
to know what data to change.

Also, fix the delete responsibilty button actually
deleting responsibilities with similar fix.

For the Ids of input elements, I just used the
randomized UUID generated for the purpose of keys